refactor(sessions): move store key ownership
