+++ src/gateway/board-store.ts
@@ -2,7 +2,10 @@ import { SqliteBoardStore } from "../boards/sqlite-board-store.js";
 import { getRuntimeConfig } from "../config/io.js";
 import { resolveStorePath } from "../config/sessions/paths.js";
 import { resolveSqliteTargetFromSessionStorePath } from "../config/sessions/session-sqlite-target.js";
-import { resolveSessionStoreAgentId, resolveSessionStoreKey } from "./session-store-key.js";
+import {
+  resolveSessionStoreAgentId,
+  resolveSessionStoreKey,
+} from "../config/sessions/session-store-key.js";
 
 export const boardStore = new SqliteBoardStore({
   resolveSession: (sessionKey) => {
