+++ src/gateway/server-methods/session-catalog-entry-snapshot.ts
@@ -8,10 +8,10 @@ import {
   listSessionEntriesReadOnly,
   type SessionEntrySummary,
 } from "../../config/sessions/session-accessor.js";
+import { resolveStoredSessionKeyForAgentStore } from "../../config/sessions/session-store-key.js";
 import type { OpenClawConfig } from "../../config/types.openclaw.js";
 import type { SessionCatalogEntrySnapshot } from "../../plugins/session-catalog.js";
 import { normalizeAgentId, resolveAgentIdFromSessionKey } from "../../routing/session-key.js";
-import { resolveStoredSessionKeyForAgentStore } from "../session-store-key.js";
 import { projectSessionActor } from "../session-utils-row.js";
 
 type SessionCatalogRequestEntrySnapshot = {
