+++ src/infra/state-migrations.session-roundtrip.test.ts
@@ -14,8 +14,8 @@ import type { OpenClawConfig } from "../config/config.js";
 import { canonicalizeMainSessionAlias } from "../config/sessions/main-session.js";
 import { resolveMainSessionKey } from "../config/sessions/main-session.js";
 import { resolveSessionKey } from "../config/sessions/session-key.js";
+import { resolveSessionStoreKey } from "../config/sessions/session-store-key.js";
 import { resolveCronAgentSessionKey } from "../cron/isolated-agent/session-key.js";
-import { resolveSessionStoreKey } from "../gateway/session-store-key.js";
 import { normalizeMainKey } from "../routing/session-key.js";
 
 function makeNonDefaultAgentCfg(overrides?: Partial<OpenClawConfig>): OpenClawConfig {
